requirements.txt
aiohttp
argparse
uvloop
numpy
pandas
geopandas
#sqlalchemy
#psycopg2
scikit-learn
aio-pika
pyyaml
#pyodbc
joblib
